+++ app/Http/Controllers/Admin/NotificationController.php
@@ -73,18 +73,13 @@ class NotificationController extends Controller
         $imageId = $notification->image_id;
         if($request->hasFile('image')){
             $image = $request->file('image');
-            $path = $image->store('all_images', 'public');
+
+            $media = resizeImage($image, $this->storagePath);
             
-            // Create media record
-            $media = Media::create([
-                'path' => $path,
-                'type' => 'image'
-            ]);
-            
-            $imageId = $media->id;
+            $imageId = $media->id ?? null;
             
             // Delete old image if exists
-            if($notification->image_id && $notification->image){
+            if($imageId && $notification->image_id && $notification->image){
                 Storage::disk('public')->delete($notification->image->path);
                 $notification->image->delete();
             }
